🧹 [code health improvement] Refactor automated-review.ts into modular files
🎯 What: Extracted types, database logic, business rules, and engine logic from automated-review.ts into a new review/ directory. The original automated-review.ts now acts as a re-export barrel file.

💡 Why: The original file was over 500 lines and mixed multiple responsibilities (DB queries, rule definitions, orchestration). This refactoring improves maintainability and readability by separating concerns.

✅ Verification: Ran pnpm lint and tests for the @settler/api package. Existing Jest ESM test failures related to uuid were verified to be pre-existing and unrelated.

✨ Result: The codebase is cleaner and follows standard architectural boundaries. Existing imports will seamlessly continue to work via the re-export.
