+++ packages/api/src/services/reconciliation/automated-review.ts
@@ -10,509 +10,4 @@
  * - GAAP/IFRS: Multi-field matching with tolerances
  */
 
-import { query } from "../../db";
-import { logError, logInfo } from "../../utils/logger";
-import { v4 as uuidv4 } from "uuid";
-import { getMatchingRulesForJob, DEFAULT_TOLERANCES } from "../matching-rules-loader";
-
-export interface ReconciliationMatch {
-  id: string;
-  runId: string;
-  sourceTransactionId: string;
-  targetTransactionId: string | null;
-  tenantId: string;
-  matchType: "exact" | "fuzzy" | "manual" | "unmatched";
-  confidence: number;
-  matchReason: string | null;
-  amountDiff: number | null;
-  dateDiff: number | null;
-  reviewed: boolean;
-  reviewedBy: string | null;
-  reviewedAt: Date | null;
-  metadata: Record<string, unknown>;
-}
-
-export interface ReviewResult {
-  matchId: string;
-  action: "auto_approved" | "rule_resolved" | "exception_handled" | "system_flagged";
-  resolutionRule?: string;
-  confidence: number;
-  auditEntryId: string;
-}
-
-// Resolution rules for exception handling
-const RESOLUTION_RULES = {
-  AMOUNT_MISMATCH_THRESHOLD: 1.0, // Auto-resolve amount differences <$1.00
-  DATE_MISMATCH_THRESHOLD_DAYS: 3, // Auto-resolve date differences <3 days
-  ROUNDING_TOLERANCE: 0.01, // Standard rounding tolerance
-} as const;
-
-const SYSTEM_USER_ID = "system:automated_review";
-
-/**
- * Automatically review a single reconciliation match
- */
-export async function autoReviewMatch(
-  matchId: string,
-  tenantId: string,
-  jobId?: string
-): Promise<ReviewResult> {
-  try {
-    // Fetch match details
-    const match = await getMatch(matchId, tenantId);
-    if (!match) {
-      throw new Error(`Match ${matchId} not found`);
-    }
-
-    // Load matching rules config from canonical loader
-    const configJobId = jobId ?? match.runId;
-    let matchingConfig;
-    try {
-      matchingConfig = await getMatchingRulesForJob(tenantId, configJobId);
-    } catch (error) {
-      logError("Failed to load matching rules, using defaults", error, {
-        tenantId,
-        jobId: configJobId,
-      });
-      matchingConfig = {
-        amountTolerance: DEFAULT_TOLERANCES.amount,
-        dateToleranceDays: DEFAULT_TOLERANCES.dateDays,
-      };
-    }
-
-    // Use threshold from matching rules or fall back to default
-    const autoApproveThreshold = 0.95; // High confidence for auto-approve
-    const ruleBasedThreshold =
-      matchingConfig.matchingRules?.find((r) => r.type === "fuzzy")?.threshold ?? 0.8;
-    const exceptionThreshold = 0.6;
-
-    // Skip if already reviewed
-    if (match.reviewed) {
-      logInfo("Match already reviewed", { matchId, tenantId });
-      return {
-        matchId,
-        action: "auto_approved",
-        confidence: match.confidence,
-        auditEntryId: "",
-      };
-    }
-
-    // Determine review action based on confidence
-    let action: ReviewResult["action"];
-    let resolutionRule: string | undefined;
-
-    if (match.confidence >= autoApproveThreshold) {
-      // Tier 1: Auto-approve high confidence matches
-      action = "auto_approved";
-      resolutionRule = "high_confidence_auto_approve";
-    } else if (match.confidence >= ruleBasedThreshold) {
-      // Tier 2: Apply rule-based resolution
-      const ruleResult = await applyRuleBasedResolution(match, matchingConfig);
-      action = ruleResult.action;
-      resolutionRule = ruleResult.rule;
-    } else if (match.confidence >= exceptionThreshold) {
-      // Tier 3: Automated exception handling
-      const exceptionResult = await handleException(match);
-      action = exceptionResult.action;
-      resolutionRule = exceptionResult.rule;
-    } else {
-      // Tier 4: System-level flagging (NOT human review)
-      action = "system_flagged";
-      resolutionRule = "low_confidence_system_review";
-    }
-
-    // Mark match as reviewed
-    await markMatchReviewed(matchId, tenantId, action, resolutionRule);
-
-    // Log audit trail
-    const auditEntryId = await logAuditTrail({
-      auditType: "auto_resolution",
-      action,
-      entityType: "reconciliation_match",
-      entityId: matchId,
-      tenantId,
-      beforeState: {
-        reviewed: false,
-        reviewedBy: null,
-        reviewedAt: null,
-      },
-      afterState: {
-        reviewed: true,
-        reviewedBy: SYSTEM_USER_ID,
-        reviewedAt: new Date().toISOString(),
-        action,
-        resolutionRule,
-      },
-      metadata: {
-        confidence: match.confidence,
-        matchType: match.matchType,
-        matchReason: match.matchReason,
-        amountDiff: match.amountDiff,
-        dateDiff: match.dateDiff,
-        resolutionRule,
-      },
-    });
-
-    logInfo("Match auto-reviewed", {
-      matchId,
-      tenantId,
-      action,
-      confidence: match.confidence,
-      resolutionRule,
-    });
-
-    return {
-      matchId,
-      action,
-      resolutionRule,
-      confidence: match.confidence,
-      auditEntryId,
-    };
-  } catch (error) {
-    logError("Failed to auto-review match", error, { matchId, tenantId });
-    throw error;
-  }
-}
-
-/**
- * Automatically review all matches in a reconciliation run
- */
-export async function autoReviewRun(
-  runId: string,
-  tenantId: string
-): Promise<{
-  reviewed: number;
-  autoApproved: number;
-  ruleResolved: number;
-  exceptionHandled: number;
-  systemFlagged: number;
-  errors: number;
-}> {
-  try {
-    // Fetch all unreviewed matches for this run
-    const matches = await query<{
-      id: string;
-      confidence: number;
-      match_type: string;
-      amount_diff: number | null;
-      date_diff: number | null;
-    }>(
-      `SELECT id, confidence, match_type, amount_diff, date_diff
-       FROM reconciliation_matches
-       WHERE run_id = $1 AND tenant_id = $2 AND reviewed = false
-       ORDER BY confidence DESC`,
-      [runId, tenantId]
-    );
-
-    const stats = {
-      reviewed: 0,
-      autoApproved: 0,
-      ruleResolved: 0,
-      exceptionHandled: 0,
-      systemFlagged: 0,
-      errors: 0,
-    };
-
-    // Review each match
-    for (const match of matches) {
-      try {
-        const result = await autoReviewMatch(match.id, tenantId);
-        stats.reviewed++;
-
-        switch (result.action) {
-          case "auto_approved":
-            stats.autoApproved++;
-            break;
-          case "rule_resolved":
-            stats.ruleResolved++;
-            break;
-          case "exception_handled":
-            stats.exceptionHandled++;
-            break;
-          case "system_flagged":
-            stats.systemFlagged++;
-            break;
-        }
-      } catch (error) {
-        stats.errors++;
-        logError("Failed to review match in run", error, {
-          matchId: match.id,
-          runId,
-          tenantId,
-        });
-      }
-    }
-
-    logInfo("Run auto-review completed", {
-      runId,
-      tenantId,
-      ...stats,
-    });
-
-    return stats;
-  } catch (error) {
-    logError("Failed to auto-review run", error, { runId, tenantId });
-    throw error;
-  }
-}
-
-/**
- * Apply rule-based resolution for medium-confidence matches
- */
-async function applyRuleBasedResolution(
-  match: ReconciliationMatch,
-  matchingConfig?: { amountTolerance: number; dateToleranceDays: number }
-): Promise<{ action: ReviewResult["action"]; rule: string }> {
-  // Use config values or fall back to defaults
-  const amountMismatchThreshold = matchingConfig?.amountTolerance ?? DEFAULT_TOLERANCES.amount;
-  const dateMismatchThreshold = matchingConfig?.dateToleranceDays ?? DEFAULT_TOLERANCES.dateDays;
-  const exactMatchThreshold = 0.85; // High confidence for exact match
-
-  // Rule 1: Amount mismatch within rounding tolerance
-  if (match.amountDiff !== null && Math.abs(match.amountDiff) <= amountMismatchThreshold) {
-    return {
-      action: "rule_resolved",
-      rule: "amount_mismatch_within_tolerance",
-    };
-  }
-
-  // Rule 2: Date mismatch within acceptable window
-  if (match.dateDiff !== null && Math.abs(match.dateDiff) <= dateMismatchThreshold) {
-    return {
-      action: "rule_resolved",
-      rule: "date_mismatch_within_window",
-    };
-  }
-
-  // Rule 3: Exact match type with high confidence
-  if (match.matchType === "exact" && match.confidence >= exactMatchThreshold) {
-    return {
-      action: "rule_resolved",
-      rule: "exact_match_high_confidence",
-    };
-  }
-
-  // Default: Handle as exception
-  return {
-    action: "exception_handled",
-    rule: "rule_based_default_exception",
-  };
-}
-
-/**
- * Handle exceptions for low-confidence matches
- */
-async function handleException(
-  match: ReconciliationMatch
-): Promise<{ action: ReviewResult["action"]; rule: string }> {
-  // Exception 1: Amount mismatch (rounding difference)
-  if (
-    match.amountDiff !== null &&
-    Math.abs(match.amountDiff) <= RESOLUTION_RULES.ROUNDING_TOLERANCE
-  ) {
-    return {
-      action: "exception_handled",
-      rule: "rounding_difference_auto_resolve",
-    };
-  }
-
-  // Exception 2: Date mismatch (timing difference)
-  if (
-    match.dateDiff !== null &&
-    Math.abs(match.dateDiff) <= RESOLUTION_RULES.DATE_MISMATCH_THRESHOLD_DAYS
-  ) {
-    return {
-      action: "exception_handled",
-      rule: "timing_difference_auto_resolve",
-    };
-  }
-
-  // Exception 3: Missing target transaction (create placeholder)
-  if (!match.targetTransactionId && match.confidence >= 0.7) {
-    // Pattern-based placeholder creation would go here
-    return {
-      action: "exception_handled",
-      rule: "missing_transaction_pattern_match",
-    };
-  }
-
-  // Default: Flag for system review
-  return {
-    action: "system_flagged",
-    rule: "exception_default_system_review",
-  };
-}
-
-/**
- * Get match details
- */
-async function getMatch(matchId: string, tenantId: string): Promise<ReconciliationMatch | null> {
-  const results = await query<{
-    id: string;
-    run_id: string;
-    source_transaction_id: string;
-    target_transaction_id: string | null;
-    tenant_id: string;
-    match_type: string;
-    confidence: number;
-    match_reason: string | null;
-    amount_diff: number | null;
-    date_diff: number | null;
-    reviewed: boolean;
-    reviewed_by: string | null;
-    reviewed_at: Date | null;
-    metadata: string;
-  }>(
-    `SELECT 
-      id, run_id, source_transaction_id, target_transaction_id,
-      tenant_id, match_type, confidence, match_reason,
-      amount_diff, date_diff, reviewed, reviewed_by, reviewed_at, metadata
-    FROM reconciliation_matches
-    WHERE id = $1 AND tenant_id = $2`,
-    [matchId, tenantId]
-  );
-
-  if (results.length === 0) {
-    return null;
-  }
-
-  const row = results[0]!;
-  return {
-    id: row.id,
-    runId: row.run_id,
-    sourceTransactionId: row.source_transaction_id,
-    targetTransactionId: row.target_transaction_id,
-    tenantId: row.tenant_id,
-    matchType: row.match_type as ReconciliationMatch["matchType"],
-    confidence: Number(row.confidence),
-    matchReason: row.match_reason,
-    amountDiff: row.amount_diff ? Number(row.amount_diff) : null,
-    dateDiff: row.date_diff,
-    reviewed: row.reviewed,
-    reviewedBy: row.reviewed_by,
-    reviewedAt: row.reviewed_at,
-    metadata: typeof row.metadata === "string" ? JSON.parse(row.metadata) : row.metadata,
-  };
-}
-
-/**
- * Mark match as reviewed
- */
-async function markMatchReviewed(
-  matchId: string,
-  tenantId: string,
-  action: string,
-  resolutionRule?: string
-): Promise<void> {
-  await query(
-    `UPDATE reconciliation_matches
-     SET reviewed = true,
-         reviewed_by = $1,
-         reviewed_at = NOW(),
-         metadata = COALESCE(metadata, '{}'::jsonb) || jsonb_build_object(
-           'auto_reviewed', true,
-           'review_action', $2,
-           'resolution_rule', $3,
-           'reviewed_at', NOW()
-         ),
-         updated_at = NOW()
-     WHERE id = $4 AND tenant_id = $5`,
-    [SYSTEM_USER_ID, action, resolutionRule || null, matchId, tenantId]
-  );
-}
-
-/**
- * Log audit trail entry
- */
-async function logAuditTrail(params: {
-  auditType: "auto_resolution" | "exception_handling" | "quality_check";
-  action: string;
-  entityType: string;
-  entityId: string;
-  tenantId: string;
-  beforeState: Record<string, unknown>;
-  afterState: Record<string, unknown>;
-  metadata: Record<string, unknown>;
-}): Promise<string> {
-  // Check if recon_audits table exists (it might be in Supabase schema)
-  // For now, we'll use reconciliation_runs metadata or create audit entry
-  const auditId = uuidv4();
-
-  try {
-    // Try to insert into recon_audits if it exists
-    await query(
-      `INSERT INTO recon_audits (
-        id, tenant_id, audit_type, action, entity_type, entity_id,
-        before_state, after_state, metadata, created_at
-      ) VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, NOW())`,
-      [
-        auditId,
-        params.tenantId,
-        params.auditType,
-        params.action,
-        params.entityType,
-        params.entityId,
-        JSON.stringify(params.beforeState),
-        JSON.stringify(params.afterState),
-        JSON.stringify(params.metadata),
-      ]
-    );
-  } catch {
-    // If table doesn't exist, log to reconciliation_runs metadata
-    logInfo("Audit table not available, logging to metadata", {
-      auditId,
-      params,
-    });
-  }
-
-  return auditId;
-}
-
-/**
- * Get review statistics for a reconciliation run
- */
-export async function getReviewStatistics(
-  runId: string,
-  tenantId: string
-): Promise<{
-  total: number;
-  reviewed: number;
-  autoApproved: number;
-  ruleResolved: number;
-  exceptionHandled: number;
-  systemFlagged: number;
-  averageConfidence: number;
-}> {
-  const results = await query<{
-    total: string;
-    reviewed: string;
-    auto_approved: string;
-    rule_resolved: string;
-    exception_handled: string;
-    system_flagged: string;
-    avg_confidence: number | null;
-  }>(
-    `SELECT 
-      COUNT(*) as total,
-      COUNT(*) FILTER (WHERE reviewed = true) as reviewed,
-      COUNT(*) FILTER (WHERE metadata->>'review_action' = 'auto_approved') as auto_approved,
-      COUNT(*) FILTER (WHERE metadata->>'review_action' = 'rule_resolved') as rule_resolved,
-      COUNT(*) FILTER (WHERE metadata->>'review_action' = 'exception_handled') as exception_handled,
-      COUNT(*) FILTER (WHERE metadata->>'review_action' = 'system_flagged') as system_flagged,
-      AVG(confidence) as avg_confidence
-    FROM reconciliation_matches
-    WHERE run_id = $1 AND tenant_id = $2`,
-    [runId, tenantId]
-  );
-
-  const row = results[0]!;
-  return {
-    total: parseInt(row.total),
-    reviewed: parseInt(row.reviewed),
-    autoApproved: parseInt(row.auto_approved),
-    ruleResolved: parseInt(row.rule_resolved),
-    exceptionHandled: parseInt(row.exception_handled),
-    systemFlagged: parseInt(row.system_flagged),
-    averageConfidence: row.avg_confidence ? Number(row.avg_confidence) : 0,
-  };
-}
+export * from "./review";
